requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Werkzeug==2.3.7
PyJWT==2.8.0
Flask-Migrate==4.0.5 # Opsional, untuk migrasi DB